refactor: consistent validator/validation
Use a consistent name for the retention validation (matching the schema
validation module).
